update streaming principle
The point was always that people should not have to request special pages to get content streamed in realtime from the server.

However, in some situations, realtime is useful for authors. For example, when playing-back their past actions, there’s no use in seeing, say, 4 hours of inactivity.

This change is harmonises these 2 points.