Add search
Initially very simple: server side compiles a list of words from blog posts and help articles, and creates a JSON file.

The word the user searches for is searched for *exactly* in the words array (this could/should be changed to use " <input>" to do a partial match.

The CSS is stupid simple, but also uses the :empty selector, which I'd rather chuffed with :)